Move miral::TestWlcsDisplayServer to the mirtest-dev package
