+++ tests/mir_test_framework/CMakeLists.txt
@@ -25,6 +25,10 @@ add_definitions(
 string (REPLACE " -flto " " " CMAKE_C_FLAGS ${CMAKE_C_FLAGS})
 string (REPLACE " -flto " " " CMAKE_CXX_FLAGS ${CMAKE_CXX_FLAGS})
 
+if (MIR_ENABLE_WLCS_TESTS)
+    set(MIR_WLCS_TEST_FRAMEWORK test_wlcs_display_server.cpp ${PROJECT_SOURCE_DIR}/include/test/miral/test_wlcs_display_server.h)
+endif()
+
 add_library(mir-public-test-framework OBJECT
   ${PROJECT_SOURCE_DIR}/include/test/mir_test_framework/any_surface.h
   any_surface.cpp
@@ -64,6 +68,7 @@ add_library(mir-public-test-framework OBJECT
   open_wrapper.cpp
   ${PROJECT_SOURCE_DIR}/tests/include/mir_test_framework/open_wrapper.h
   test_server.cpp  ${PROJECT_SOURCE_DIR}/include/test/miral/test_server.h
+  ${MIR_WLCS_TEST_FRAMEWORK}
 )
 
 set_property(
